tests/test_requirements.txt
# Test dependencies for Blockchain Payment Processor

# Core testing framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Additional pytest plugins
pytest-xdist==3.4.0          # Parallel testing
pytest-html==4.1.1           # HTML test reports
pytest-benchmark==4.0.0      # Performance benchmarking
pytest-watch==4.2.0          # Auto-rerun tests on file changes
pytest-sugar==0.9.7          # Better test output formatting
pytest-clarity==1.0.1        # Better assertion failure messages

# Mocking and test utilities
responses==0.24.1             # Mock HTTP requests
freezegun==1.2.2              # Mock datetime
factory-boy==3.3.0            # Test data generation
faker==20.1.0                 # Fake data generation

# Code quality tools
black==23.11.0                # Code formatting
isort==5.12.0                 # Import sorting
flake8==6.1.0                 # Linting
mypy==1.7.1                   # Type checking
bandit==1.7.5                 # Security linting

# Coverage tools
coverage[toml]==7.3.2         # Coverage measurement
pytest-cov==4.1.0             # Pytest coverage plugin

# Database testing
pytest-postgresql==5.0.0     # PostgreSQL test fixtures (optional)
alembic==1.12.1               # Database migrations (already in main requirements)

# Performance testing
memory-profiler==0.61.0      # Memory usage profiling
line-profiler==4.1.1         # Line-by-line profiling

# API testing helpers
httpx==0.25.2                 # Already in main requirements - async HTTP client

# Documentation testing
pytest-doctestplus==1.0.0    # Enhanced doctest support

# Snapshot testing
syrupy==4.6.1                # Snapshot testing for API responses

# Load testing (optional)
locust==2.17.0                # Load testing tool

# Test environment management
python-dotenv==1.0.0          # Already in main requirements
pytest-env==1.1.3             # Environment variable management for tests

# Debugging tools
pytest-pdb==0.2.0             # PDB integration
pdbpp==0.10.3                 # Enhanced PDB

# Test reporting
pytest-json-report==1.5.0    # JSON test reports
allure-pytest==2.13.2        # Allure test reporting

# Security testing
safety==2.3.5                 # Check for known security vulnerabilities

# Development tools
pre-commit==3.5.0             # Git hooks for code quality
commitizen==3.12.0            # Conventional commits

# Type checking
types-requests==2.31.0.10     # Type stubs for requests
types-python-dateutil==2.8.19.14

# Test fixtures and utilities
pytest-factoryboy==2.6.0     # Factory Boy integration
pytest-lazy-fixture==0.6.3   # Lazy fixture evaluation

# Database tools for testing
sqlalchemy-utils==0.41.1     # Database utilities for testing

# Async testing helpers
pytest-trio==0.8.0            # Trio async testing support (if needed)
pytest-tornado==0.8.1        # Tornado testing support (if needed)

# Mock external services
betamax==0.8.1                # HTTP request/response recording
vcrpy==5.1.0                  # Record and replay HTTP interactions